agent-core/requirements.txt
langchain==0.1.0
langchain-openai==0.0.2
langchain-google-genai==0.0.5
langchain-community==0.0.10
kubernetes==28.1.0
pydantic==2.5.3
python-dotenv==1.0.0
pyyaml==6.0.1
